sdk/pyproject.toml
[project]
name = "sdk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-ai-ml>=1.23.1",
    "azure-identity>=1.19.0",
    "click>=8.1.8",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "python-dotenv>=1.0.1",
    "ruff>=0.8.6",
]
